Added initial support for LSU.

+++ readConfig.py
@@ -53,15 +53,15 @@ class ReadConfig(object):
       if option == 'name':
         sampleParams[option] = data
       elif option == 'edit_dist':
-        sampleParams['edit_dist'] = int(data)
+        sampleParams['edit_dist'] = float(data)
       elif option == 'min_align_len':
-        sampleParams['min_align_len'] = int(data)
+        sampleParams['min_align_len'] = float(data)
       else:
         if data != '':
           sampleParams[option] = [x.strip() for x in data.split(',')]
         else:
           sampleParams[option] = []
-          
+
     return sampleParams
 
   def validateMothurCompatibility(self, path):
